refactor(frontend): merge util.ts into utils.ts
