+++ frontend/src/lib/utils.ts
@@ -1,5 +1,16 @@
 import { resolve } from '$app/paths';
 
+/** Normalize text for search: strip diacritics, punctuation, and collapse whitespace. */
+export function normalizeText(s: string): string {
+	return s
+		.normalize('NFD')
+		.replace(/[\u0300-\u036f]/g, '') // strip diacritics
+		.replace(/[^\w\s]/g, '') // strip punctuation
+		.replace(/\s+/g, ' ') // collapse whitespace
+		.trim()
+		.toLowerCase();
+}
+
 /** Wrapper around resolve() that accepts a plain string (for dynamic URLs). */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const resolveHref = (url: string) => resolve(url as any);
